Refactored relation property endpoints
