+++ charmClient/charmClient.ts
@@ -6,12 +6,9 @@ import type {
   InviteLink,
   Page,
   PaymentMethod,
-  Space,
   TelegramUser,
-  TokenGateToRole,
   User,
-  UserDetails,
-  UserWallet
+  UserDetails
 } from '@charmverse/core/prisma';
 import type { FiatCurrency, IPairQuote } from 'connectors/chains';
 
@@ -22,7 +19,6 @@ import type { BlockPatch, Block as FBBlock } from 'lib/focalboard/block';
 import type { InviteLinkPopulated } from 'lib/invites/getInviteLink';
 import type { PublicInviteLinkRequest } from 'lib/invites/getPublicInviteLink';
 import type { InviteLinkWithRoles } from 'lib/invites/getSpaceInviteLinks';
-import type { Web3LoginRequest } from 'lib/middleware/requireWalletSignature';
 import type { CreateEventPayload } from 'lib/notifications/interfaces';
 import type { FailedImportsError } from 'lib/notion/types';
 import type { ModifyChildPagesResponse, PageLink } from 'lib/pages';
@@ -32,7 +28,7 @@ import type { ITokenMetadata, ITokenMetadataRequest } from 'lib/tokens/tokenData
 import { encodeFilename } from 'lib/utilities/encodeFilename';
 import type { SocketAuthResponse } from 'lib/websockets/interfaces';
 import type { LoggedInUser } from 'models';
-import type { SyncRelationPropertyPayload } from 'pages/api/blocks/sync-relation-property';
+import type { SyncRelationPropertyPayload } from 'pages/api/blocks/relation/sync';
 import type { SyncRelationPropertyValuePayload } from 'pages/api/blocks/sync-relation-property-value';
 import type { ImportGuildRolesPayload } from 'pages/api/guild-xyz/importRoles';
 import type { TelegramAccount } from 'pages/api/telegram/connect';
