+++ components/common/BoardEditor/focalboard/src/components/table/tableHeaders.tsx
@@ -200,7 +200,6 @@ function TableHeaders(props: Props): JSX.Element {
             if (relationData?.showOnRelatedBoard) {
               syncRelationProperty({
                 boardId: board.id,
-                action: 'create',
                 templateId: template.id
               });
             }
